refactor: extract humanize_bytes/1 to LogflareWeb.Utils
