--- test/logflare_web/components/query_components_test.exs
@@ -1,5 +1,0 @@
-defmodule LogflareWeb.QueryComponentsTest do
-  use ExUnit.Case, async: true
-
-  doctest LogflareWeb.QueryComponents, import: true
-end
